[raster] Signedness fixes.
* src/raster/ftraster.c, src/raster/ftrend1.c: Apply.

+++ src/raster/ftrend1.c
@@ -192,7 +192,7 @@
 
     bitmap->width = width;
     bitmap->rows  = height;
-    bitmap->pitch = pitch;
+    bitmap->pitch = (int)pitch;
 
     if ( FT_ALLOC_MULT( bitmap->buffer, pitch, height ) )
       goto Exit;
